chore(frontend): tidy colors and document theming
