+++ frontend/docs/theming.md
@@ -1,0 +1,19 @@
+# Темизация и плагины
+
+Проект использует Tailwind CSS с расширенной цветовой палитрой и кастомным плагином.
+
+## Цветовая система
+- Палитры `primary`, `secondary` и `accent` описаны в `tailwind.config.js`.
+- Для состояний интерфейса определены палитры `error`, `warning`, `success` и `info`.
+- Цвета связаны с CSS‑переменными в `src/index.css`, что позволяет переключать светлую и тёмную темы.
+- Неиспользуемые токены `border` и `muted` удалены для соответствия фактическому UI.
+
+## Переключение темы
+- Переменные `--primary`, `--secondary` и `--accent` задаются в `:root` и в `.dark` в `src/index.css`.
+- Модуль `src/theme` содержит функции `initializeTheme`, `toggleTheme` и другие утилиты для хранения и применения выбранной темы.
+
+## Tailwind‑плагин
+- В `tailwind.config.js` определён плагин, добавляющий утилиту `.loaded`. Класс используется, например, в компоненте `LazyImage` для плавного появления изображения после загрузки.
+
+## Конфигурация компонентов
+- Повторяющиеся классы цветов для `Input` вынесены в `src/components/ui/Input/Input.config.ts`, что упрощает поддержку и расширение компонентов.
