+++ frontend/src/components/ui/Input/index.ts
@@ -1,2 +1,2 @@
 export { Input } from './Input';
-export type { InputProps, InputSize, InputVariant, InputColor } from './Input';
+export type { InputProps, InputSize, InputVariant, InputColor } from './Input.types';
